requirements.txt
streamlit>=1.33.0
pandas>=1.5.0
Pillow>=9.0.0
streamlit-autorefresh>=0.0.3